Add: next js app template

+++ apps/pi-zero-station/camera-shot.sh
@@ -1,14 +1,30 @@
-#!/bin/sh
+#!/bin/bash
 
 # To copy to raspberry pi:
 # scp camera-shot.sh christopher@raspberrypi:/home/christopher
+# scp .env christopher@raspberrypi:/home/christopher
 
-picam_enabled=true;
-webcam_enabled=true;
-user='christopher';
-home_path='/home/christopher';
-remote_server='master';
-picam_remote_path='/shared-volume/time-lapse/picam';
+# Add to rc.local file
+# sudo vim /etc/rc.local
+# > /bin/bash /home/christopher/camera-shot.sh
+
+# Grant exec privileges
+# sudo chmod +x /etc/rc.local
+
+home_path="/home/christopher";
+remote_path="christopher@master:/shared-volume/time-lapse";
+mkdir -p "$home_path/logs";
+
+source "$home_path/.env"
+cat "$home_path/.env"
+
+# [[ $picam_enabled ]] && picam_enabled=true || picam_enabled=false;
+# webcam_enabled=false;
+# if [[ $webcam_enabled ]]; then
+#   webcam_enabled=true;
+# fi
+# sunrise=`expr $SUNRISE + 0`;
+# sunset=`expr $sunset + 0`;
 
 if $picam_enabled; then
   mkdir -p "$home_path/picam";
@@ -20,37 +36,58 @@ if $webcam_enabled; then
   # ssh -i /home/christopher/.ssh/id_ed25519 christopher@master 'mkdir -p /shared-volume/time-lapse/webcam'
 fi
 
-while true; do 
-  # File name
-  date=$(date '+%Y-%m-%d %H-%M-%S')
-  file_name="picture $date.jpg"
+while true; do
+  source "$home_path/.env"
 
-  # Raspberry PI Camera
-  if $picam_enabled; then
-    raspistill -o "$home_path/picam/$file_name" -w 2592 -h 1944 -q 100 -sh 90 -rot 90
-    scp -i /home/christopher/.ssh/id_ed25519 "$home_path/picam/$file_name" christopher@master:/shared-volume/time-lapse/picam
-    rm "$home_path/picam/$file_name"
-    echo "[picam] Picture saved: picture $date.jpg"
+  logs_file_name="$home_path/logs/`date "+%Y-%m-%d"`.txt";
+  touch $logs_file_name;
+  # current_hour
+  current_hour=`date "+%H"`;
+  # cast string to int
+  current_hour=`expr $current_hour + 0`;
+
+  date=$(date '+%Y-%m-%d %H-%M-%S');
+
+  if [[ $current_hour -ge $sunrise && $current_hour -le $sunset ]]; then
+    echo "===== Photos are enabled: $date. =====";
+
+    # File name
+    file_name="picture $date.jpg";
+
+    # Raspberry PI Camera
+    if $picam_enabled; then
+      raspistill -o "$home_path/picam/$file_name" -w $picam_width -h $picam_height -q 100 -sh 90 -rot $picam_rotation;
+      scp -i "$home_path/.ssh/id_ed25519" "$home_path/picam/$file_name" "$remote_path/picam";
+      rm "$home_path/picam/$file_name";
+      echo "[picam] Picture saved: picture $date.jpg" >> $logs_file_name;
+    fi
+
+    # Webcam
+    if $webcam_enabled; then
+      v4l2-ctl -d /dev/video1 -c white_balance_automatic=0;
+      v4l2-ctl -d /dev/video1 -c brightness=$webcam_brightness;
+      fswebcam \
+        -d $webcam_source \
+        -r $webcam_resolution \
+        -S $webcam_frames_skiped \
+        --quiet \
+        --jpeg 95 \
+        --rotate $webcam_rotation \
+        --no-banner "$home_path/webcam/$file_name";
+      scp -i "$home_path/.ssh/id_ed25519" "$home_path/webcam/$file_name" "$remote_path/webcam";
+      rm "$home_path/webcam/$file_name";
+      echo "[webcam] Picture saved: picture $date.jpg" >> $logs_file_name;
+    fi
+  else
+    echo "===== Night time, photos are disabled: $date. =====" >> $logs_file_name;
   fi
+  echo "  -Sleep time: $sleep_time seconds." >> $logs_file_name;
+  echo "  -Sunrise: $sunrise hrs. Sunset: $sunset hrs. Current: $current_hour hrs." >> $logs_file_name;
+  echo "  -Webcam enabled: $webcam_enabled, brightness: $webcam_brightness" >> $logs_file_name;
+  echo "  -Picam enabled: $picam_enabled" >> $logs_file_name;
+  scp -i "$home_path/.ssh/id_ed25519" $logs_file_name "$remote_path/logs";
 
-  # Webcam
-  if $webcam_enabled; then
-    v4l2-ctl -d /dev/video1 -c white_balance_automatic=0
-    v4l2-ctl -d /dev/video1 -c brightness=90
-    fswebcam \
-      -d /dev/video1 \
-      -r 1920x1080 \
-      -S 1 \
-      --quiet \
-      --jpeg 95 \
-      --rotate 180 \
-      --no-banner "$home_path/webcam/$file_name"
-    scp -i /home/christopher/.ssh/id_ed25519 "$home_path/webcam/$file_name" christopher@master:/shared-volume/time-lapse/webcam
-    rm "$home_path/webcam/$file_name"
-    echo "[webcam] Picture saved: picture $date.jpg"
-  fi
-
-  sleep 10
+  sleep $sleep_time
 done
 
 exit 0
